Create FileOperationListItem
