+++ shared/i18n/locales/en_US/translation.json
@@ -466,6 +466,11 @@
   "No documents found for your search filters. <1></1>": "No documents found for your search filters. <1></1>",
   "Create a new document?": "Create a new document?",
   "Clear filters": "Clear filters",
+  "Processing": "Processing",
+  "Expired": "Expired",
+  "Error": "Error",
+  "All collections": "All collections",
+  "{{userName}} requested": "{{userName}} requested",
   "Last active": "Last active",
   "Role": "Role",
   "Viewer": "Viewer",
@@ -489,10 +494,6 @@
   "Groups can be used to organize and manage the people on your team.": "Groups can be used to organize and manage the people on your team.",
   "All groups": "All groups",
   "No groups have been created yet": "No groups have been created yet",
-  "Processing": "Processing",
-  "Complete": "Complete",
-  "Expired": "Expired",
-  "Error": "Error",
   "Import started": "Import started",
   "Export in progress…": "Export in progress…",
   "It is possible to import a zip file of folders and Markdown files previously exported from an Outline instance. Support will soon be added for importing from other services.": "It is possible to import a zip file of folders and Markdown files previously exported from an Outline instance. Support will soon be added for importing from other services.",
@@ -507,8 +508,6 @@
   "Requesting Export": "Requesting Export",
   "Export Data": "Export Data",
   "Recent exports": "Recent exports",
-  "All collections": "All collections",
-  "{{userName}} requested": "{{userName}} requested",
   "Document published": "Document published",
   "Receive a notification whenever a new document is published": "Receive a notification whenever a new document is published",
   "Document updated": "Document updated",
